2006-01-25  Ales Hakl  <ales at hakl dot net>
+	* src/dfsch.c (dfsch_vector_2_list): Fixed segfaults in list->vector
+
 	* New way of handling objects in memory 
 
 2006-01-23  Ales Hakl  <ales at hakl dot net>


git-svn-id: svn://mahoro/dfsch/trunk@177 876e91b9-e3f0-0310-bb80-9da7c39ee78d

+++ dfsch/dfsch.h
@@ -242,7 +242,7 @@ extern "C" {
   /**
    * Return number of items in given list
    */
-  extern int dfsch_list_length(dfsch_object_t* list);
+  extern long dfsch_list_length(dfsch_object_t* list);
 
   /**
    * Returns given item of list.
